Feat : 식물 성장 저장&로드
- 식물의 위치, 성장 단계, 성장 시간, 타일의 상태 등 개별 인스턴스 정보를 저장하기 위해PlantSaveData 클래스 생성
- 식물 데이터(PlantData) 뿐만 아니라 내가 씨를 뿌린 그 식물에 현재 상태도 저장
- 로드한 데이터를 기반으로 SetTile

+++ Assets/Scripts/ScriptableObject/PlantData.cs
@@ -7,5 +7,5 @@ public class PlantData : ScriptableObject
     public string plantName = "Plant Name";
     public GameObject plantPrefab;
     public Tile[] growthStagesTiles;
-    public float[] growthTimes;
+    public int[] growthTimes;
 }
